requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
filelock==3.13.1
Flask==3.0.1
fsspec==2023.12.2
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
llvmlite==0.41.1
MarkupSafe==2.1.4
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.3
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
regex==2023.12.25
requests==2.31.0
sympy==1.12
tiktoken==0.5.2
torch==2.1.2
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
zipp==3.17.0
